requirement.txt
absl-py==0.7.1
astor==0.7.1
audioread==2.1.6
decorator==4.4.0
enum34==1.1.6
funcsigs==1.0.2
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
librosa==0.6.3
lightgbm==2.2.3
llvmlite==0.28.0
Markdown==3.1
mock==2.0.0
numba==0.43.1
numpy==1.16.3
pandas==0.24.2
pbr==5.2.0
pkg-resources==0.0.0
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
resampy==0.2.1
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
Werkzeug==0.15.2
